--- a/frappe/core/doctype/navbar_settings/navbar_settings.py
+++ b/frappe/core/doctype/navbar_settings/navbar_settings.py
@@ -24,8 +24,10 @@ class NavbarSettings(Document):
 
 
 def get_app_logo():
-	app_logo = frappe.get_website_settings("app_logo") or frappe.db.get_single_value(
-		"Navbar Settings", "app_logo", cache=True
+	app_logo = frappe.get_website_settings("app_logo") or frappe.get_cached_value(
+		"Navbar Settings",
+		"Navbar Settings",
+		"app_logo",
 	)
 
 	if not app_logo: