This small change speeds up *most* on-disk web template-based workloads by at least 2x because they aren't parsed on every request anymore.

My goal was just to optimize the login page render:

| Benchmark                      | before  | after                 |
|--------------------------------|:-------:|:---------------------:|
| web_requests_login_page_render | 85.1 ms | 11.8 ms: 7.23x faster |



TODO:
- [x] Don't cache things in developer mode, does it matter? Since we reload code anyway? :thinking: 
- [x] Add test for request specific jenv globals
- [ ] Test apps like `wiki` (in prod :moyai: )

closes https://github.com/frappe/caffeine/issues/29